chore: Move TwoFa and Credentials sections into its own components
